completed trainer suggested comments
